Remove _alchemical_pair_force.py file
